# Benchmark and speed up group validation.