--- a/benchmark/src/main/scala/mesosphere/marathon/state/RootGroupBenchmark.scala
+++ b/benchmark/src/main/scala/mesosphere/marathon/state/RootGroupBenchmark.scala
@@ -25,8 +25,7 @@ class GroupBenchmark {
         Container.Docker(Nil, "alpine", List(Container.PortMapping(2015, Some(0), 10000, "tcp", Some("thing")))))
     )
 
-  //@Param(value = Array("2", "10", "100", "1000"))
-  @Param(value = Array("2", "10", "100"))
+  @Param(value = Array("2", "10", "100", "1000"))
   var appsPerGroup: Int = _
   lazy val appIds = 0 until appsPerGroup