--- a/src/main/scala/mesosphere/marathon/state/Group.scala
+++ b/src/main/scala/mesosphere/marathon/state/Group.scala
@@ -10,8 +10,8 @@ import mesosphere.marathon.api.v2.Validation._
 import mesosphere.marathon.api.v2.validation.AppValidation
 import mesosphere.marathon.core.pod.PodDefinition
 import mesosphere.marathon.plugin.{ Group => IGroup }
-import mesosphere.marathon.state.Group.{ defaultApps, defaultPods, defaultGroups, defaultDependencies, defaultVersion }
-import mesosphere.marathon.state.PathId.{ validPathWithBase, StringPathId }
+import mesosphere.marathon.state.Group.{ defaultApps, defaultDependencies, defaultGroups, defaultPods, defaultVersion }
+import mesosphere.marathon.state.PathId.{ StringPathId, validPathWithBase }
 
 class Group(
     val id: PathId,
@@ -154,23 +154,57 @@ object Group extends StrictLogging {
     validator[Group] { group =>
       group.id is validPathWithBase(base)
 
-      group.transitiveApps as "apps" is every(
-        validator[AppDefinition] { app =>
-          group.group(app.id.parent) match {
-            case None => Failure
-            case Some(parentGroup) => validPathWithBase(parentGroup.id).apply(app.id)
-          }
-        })
+      group.transitiveApps as "apps" is everyApp(
+        AppDefinition.validBasicAppDefinition(enabledFeatures) and isChildOfParentId(group)
+      )
 
       group is noAppsAndPodsWithSameId
       group is noAppsAndGroupsWithSameName
       group is noPodsAndGroupsWithSameName
-      //group.groupsById.values as "groups" is every(validGroup(group.id.canonicalPath(base), enabledFeatures))
+      
       group.transitiveGroupValues as "groups" is every(
-        // TODO: Validate path with base
         noAppsAndPodsWithSameId and noAppsAndGroupsWithSameName and noPodsAndGroupsWithSameName
+          and isTrue("Group has to be child of groups with parent id") { childGroup =>
+            if (childGroup.id.parent.isRoot) group.groupsById.contains(childGroup.id)
+            else {
+              group.group(childGroup.id.parent) match {
+                case None => false
+                case Some(parentGroup) => parentGroup.groupsById.contains(childGroup.id)
+              }
+            }
+          }
       )
     }
+
+  implicit def everyApp(validator: Validator[AppDefinition]): Validator[Iterable[AppDefinition]] = {
+    new Validator[Iterable[AppDefinition]] {
+      override def apply(seq: Iterable[AppDefinition]): Result = {
+        seq.foldLeft[Result](Success) {
+          case (accum, (app)) =>
+            validator(app) match {
+              case Success => accum
+              case Failure(violations) =>
+                val scopedViolations = violations.map { violation =>
+                  violation.withPath(Descriptions.Explicit(app.id.toString))
+                }
+                accum.and(Failure(scopedViolations))
+            }
+        }
+      }
+    }
+  }
+
+  private def isChildOfParentId(group: Group): Validator[AppDefinition] = {
+    isTrue("App has to be child of group with parent id") { app =>
+      if (app.id.parent.isRoot) group.apps.contains(app.id)
+      else {
+        group.group(app.id.parent) match {
+          case None => false
+          case Some(parentGroup) => parentGroup.apps.contains(app.id)
+        }
+      }
+    }
+  }
 
   private def noAppsAndPodsWithSameId: Validator[Group] =
     isTrue("Applications and Pods may not share the same id") { group =>
@@ -179,8 +213,8 @@ object Group extends StrictLogging {
 
   private def noAppsAndGroupsWithSameName: Validator[Group] =
     isTrue("Groups and Applications may not have the same identifier.") { group =>
-      val groupIds = group.groupsById.keys
-      val clashingIds = groupIds.filter(group.apps.contains(_))
+      val childGroupIds = group.groupsById.keys
+      val clashingIds = childGroupIds.filter(group.apps.contains(_))
       if (clashingIds.nonEmpty)
         logger.info(s"Found the following clashingIds in group ${group.id}: ${clashingIds}")
       clashingIds.isEmpty
@@ -188,8 +222,8 @@ object Group extends StrictLogging {
 
   private def noPodsAndGroupsWithSameName: Validator[Group] =
     isTrue("Groups and Pods may not have the same identifier.") { group =>
-      val groupIds = group.groupsById.keys
-      val clashingIds = groupIds.filter(group.pods.contains(_))
+      val childGroupIds = group.groupsById.keys
+      val clashingIds = childGroupIds.filter(group.pods.contains(_))
       if (clashingIds.nonEmpty)
         logger.info(s"Found the following clashingIds in group ${group.id}: ${clashingIds}")
       clashingIds.isEmpty