--- a/src/test/scala/mesosphere/marathon/state/RootGroupTest.scala
+++ b/src/test/scala/mesosphere/marathon/state/RootGroupTest.scala
@@ -491,7 +491,7 @@ class RootGroupTest extends UnitTest with GroupCreation {
 
     "relative dependencies should be resolvable" in {
       Given("a group with an app having relative dependency")
-      val app1 = AppDefinition("app1".toPath, cmd = Some("foo"))
+      val app1 = AppDefinition("/group/app1".toPath, cmd = Some("foo"))
       val app2 = AppDefinition("/group/subgroup/app2".toPath, cmd = Some("bar"), dependencies = Set("../app1".toPath))
       val rootGroup = createRootGroup(groups = Set(
         createGroup("/group".toPath, apps = Map(app1.id -> app1),