Summary:
Traversing the groups recursively is slower than going over the flattened id.

This also includes a bug fix. The validation of
```
app.id is PathId.validPathWithBase(base)
```
as the based base was always the parent. The same holds up for
```
path is childOf(path.parent)
```

The invalid app definitions in the unit tests have been fixed as well.

Additionally this change introduces more expressive violations paths. See the unit test changes
for details.

JIRA issues: MARATHON_EE-883