@jeschkies 2018-01-05
# Benchmark `validateRootGroup`

sbt "benchmark/jmh:run -i 10 -wi 3 -f1 -t1 .*RootGroupBenchmark.validateRootGroup"

## Master
|(appsPerGroup)|(groupDepth) | Mode | Cnt | Score | Error | Units
|---:|---:|---:|---:|---:|----:|---|
|   2|  5| avgt| 10|   0.406| ± 0.040|ms/op
|   2| 10| avgt| 10|   0.782| ± 0.056|ms/op
|   2| 20| avgt| 10|   1.844| ± 0.150|ms/op
|  10|  5| avgt| 10|   1.677| ± 0.090|ms/op
|  10| 10| avgt| 10|   4.034| ± 0.294|ms/op
|  10| 20| avgt| 10|   8.236| ± 0.650|ms/op
| 100|  5| avgt| 10|  17.563| ± 1.376|ms/op
| 100| 10| avgt| 10|  37.673| ± 2.930|ms/op
| 100| 20| avgt| 10|  86.548| ± 4.552|ms/op
|1000|  5| avgt| 10| 179.341| ± 9.539|ms/op
|1000| 10| avgt| 10| 355.218| ±12.458|ms/op
|1000| 20| avgt| 10| 941.220| ±31.937|ms/op

## This Patch
|(appsPerGroup)|(groupDepth) | Mode | Cnt | Score | Error | Units
|---:|---:|---:|---:|---:|----:|---|
|   2|  5| avgt| 10|   0.333| ± 0.038|ms/op
|   2| 10| avgt| 10|   0.699| ± 0.207|ms/op
|   2| 20| avgt| 10|   1.284| ± 0.137|ms/op
|  10|  5| avgt| 10|   1.453| ± 0.210|ms/op
|  10| 10| avgt| 10|   3.079| ± 0.277|ms/op
|  10| 20| avgt| 10|   5.936| ± 0.468|ms/op
| 100|  5| avgt| 10|  14.520| ± 0.872|ms/op
| 100| 10| avgt| 10|  30.836| ± 2.073|ms/op
| 100| 20| avgt| 10|  60.739| ± 3.540|ms/op
|1000|  5| avgt| 10| 152.573| ± 9.179|ms/op
|1000| 10| avgt| 10| 320.610| ±22.012|ms/op
|1000| 20| avgt| 10| 631.478| ±60.635|ms/op

That is roughly a 30% speed up for 20,000 apps distributed over twenty groups.